requirements.txt
pytest==5.2.1
selenium==3.141.0
webdriver_manager==1.8.2
requests==2.22.0
Pillow==6.2.0